+++ packages/alfa-rules/src/common/dom/get-positioning-parent.ts
@@ -2,7 +2,8 @@ import { Device } from "@siteimprove/alfa-device";
 import { Element } from "@siteimprove/alfa-dom";
 import { Option } from "@siteimprove/alfa-option";
 
-import { isPositioned } from "../predicate";
+import { isPositioned } from "../predicate/is-positioned";
+
 import { getOffsetParent } from "./get-offset-parent";
 
 const { isElement } = Element;
