+++ packages/alfa-rules/tsconfig.json
@@ -13,10 +13,12 @@
     "src/common/applicability/audio.ts",
     "src/common/applicability/video.ts",
     "src/common/diagnostic/contrast.ts",
+    "src/common/dom/get-colors.ts",
     "src/common/dom/get-nodes-between.ts",
     "src/common/dom/get-offset-parent.ts",
     "src/common/dom/get-positioning-parent.ts",
     "src/common/dom/lowest-common-ancestor.ts",
+    "src/common/expectation/contrast.ts",
     "src/common/expectation/get-refresh-time.ts",
     "src/common/expectation/media-text-alternative.ts",
     "src/common/expectation/video-description-track-accurate.ts",
@@ -76,8 +78,6 @@
     "src/common/predicate/is-whitespace.ts",
     "src/common/predicate/is-wide-enough.ts",
     "src/common/predicate/reference-same-resource.ts",
-    "src/common/style/contrast.ts",
-    "src/common/style/get-colors.ts",
     "src/experimental.ts",
     "src/index.ts",
     "src/rules.ts",
